==> Notebooks/02_cleaning.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns


df = pd.read_csv("/Users/rathsusan/Documents/Applied-Machine-Learning-Project/Data/Raw/online_shoppers_intention 2.csv")
df.head()

df.isnull().sum

df.dtypes


df = df.drop(columns=['Month'])

df.head()


df = df.drop(columns=['PageValues'])
df.head()

categorical_cols = ['OperatingSystems', 'Browser', 'Region', 'TrafficType', 'VisitorType']

# Apply one-hot encoding
df = pd.get_dummies(df, columns=categorical_cols, drop_first=True)

df.columns

df.dtypes

output_path = "../Data/Processed/online_shoppers_clean.csv"
df.to_csv(output_path, index=False)

print(f"Cleaned dataset saved to: {output_path}")

check_df = pd.read_csv(output_path)
print("First five rows of the saved dataset:")
display(check_df.head())
