==> Notebooks/03_train_test_split.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split


df = pd.read_csv("../Data/Processed/online_shoppers_clean.csv")


X = df.drop("Revenue", axis=1)
y = df["Revenue"]


X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, stratify=y, random_state=42
)


print("X_train:", X_train.shape)
print("X_test:", X_test.shape)

# just double checking if it worked
print("y_train mean:", y_train.mean().round(3))
print("y_test mean:", y_test.mean().round(3))



X_train.to_csv("../Data/Processed/X_train.csv", index=False)
X_test.to_csv("../Data/Processed/X_test.csv", index=False)
y_train.to_csv("../Data/Processed/y_train.csv", index=False)
y_test.to_csv("../Data/Processed/y_test.csv", index=False)
